pyproject.toml
[project]
name = "lane-detection-student"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
